Fix meta header parsing
I was trying to be clever and save parsing effort by assuming the server would send the header as a separate message from the body, and only scan backwards from the buffer we get in `handleInput(framer:)`.
Turns out that's not the case and some servers just send everything as one blob. So instead we do the simple thing and scan forward until we hit <CR><LF>